+++ app/src/main/java/com/example/gogreen/data/StartChargingResponse.kt
@@ -6,8 +6,8 @@ data class StartChargingResponse(
     val energyConsumed: Double? = 0.0,
     val price: Double? = 0.0,
     val requiredEnergy: Double? = 0.0,
-    val timePassed: Double? = 0.0,
-    val totalTime: Double? = 0.0,
-    val transferRate: Double? = 0.0,
+    val timePassed: Int? = 0,
+    val totalTime: Int? = 0,
+    val transferRate: Int? = 0,
     val totalPrice: Double? = 0.0
 )
